+++ Example/MultiSetWidget/TestItem.cpp
@@ -1,0 +1,6 @@
+#include "TestItem.h"
+
+GENERATE_SOURCE_STANDARDCLASS(Note, NoteRecord)
+
+
+GENERATE_SOURCE_STANDARDSET(MetaNoteSet, "Notes", "Note")