+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Ramio changelog
 
+## 0.1.5f4 MultiSetWidget [15.12.2020]
+- Новый пример использования MultiSetWidget
+- Подправлены наборы MergerItemSet и MultiMergerItemSet
+- Добавлена обработка для новых полей Type, State, Flags в MetaItemsModel
+
 ## 0.1.5f3 fieldValueToString & differingDataFields [14.12.2020]
 - Функция поиска различающихся полей differingDataFields
 - Метод valueToString структуры MetaItemData для вывода значения полей по индексу